Major changes for QueryBuilder and FieldMap!
The StagedQueryBuilder proofed to be too complex. Issues where difficult to debug.
It has been replaced by more a concrete FilterQueryBuilder and UpdateQueryBuilder.
The FilterQueryBuilder directly calls the filter parser. This is no longer part of
 the prepare step. The of this isn't an iterable with arrays as keys, but a set of
 FilterItem objects.
Update instructions are already given as UpdateInstruction objects, so no parser
 is required.
Dropped the compose step, the logic (previous build) is selected and applied by
 the query builder. Custom filter and update logic is registered at the query
 builder for a specific field or operator.
Custom filer/update logic may call `$next` to apply the default logic.
Prepare and finalize are a single callable. There are helper objects which take
 multiple arguments, and thus work similar to previous steps in staged builder.

The field map is no longer invokable. Instead it has specific method for apply the
 map to a filter, update instruction, items (for save) or a result.
Field maps don't need to be flipped, the have both a normal and inverse copy of
 the map. If a field is duplicated, an exception is thrown.
Implemented ConfiguredFieldMap::__set_state().

The library uses Jasny\Immutable for `with` methods and to disable dynamic
 properties.